Moved and changed the name ConsoleOutput to logicConsoleOutput.py in the gui.controller file.
