+++ gui/controller/logicConsoleOutput.py
